UHF-9961: Enable helfi_annif module, install elasticsearch index
