+++ conf/cmi/core.entity_view_display.node.news_item.text_converter.yml
@@ -1,0 +1,67 @@
+uuid: 48105040-80f2-4420-9b3c-56ba665333b0
+langcode: en
+status: true
+dependencies:
+  config:
+    - core.entity_view_mode.node.text_converter
+    - field.field.node.news_item.field_content
+    - field.field.node.news_item.field_lead_in
+    - field.field.node.news_item.field_main_image
+    - field.field.node.news_item.field_main_image_caption
+    - field.field.node.news_item.field_news_groups
+    - field.field.node.news_item.field_news_item_links_link
+    - field.field.node.news_item.field_news_item_links_title
+    - field.field.node.news_item.field_news_item_tags
+    - field.field.node.news_item.field_news_neighbourhoods
+    - field.field.node.news_item.field_radioactivity
+    - field.field.node.news_item.field_short_title
+    - node.type.news_item
+  module:
+    - entity_reference_revisions
+    - helfi_platform_config
+    - user
+_core:
+  default_config_hash: fHNnwg186A981K7d578If1MCkuYBmQqaxabpDSAZwIc
+id: node.news_item.text_converter
+targetEntityType: node
+bundle: news_item
+mode: text_converter
+content:
+  field_content:
+    type: entity_reference_revisions_entity_view
+    label: hidden
+    settings:
+      view_mode: default
+      link: ''
+    third_party_settings: {  }
+    weight: 2
+    region: content
+  field_lead_in:
+    type: text_span
+    label: hidden
+    settings: {  }
+    third_party_settings: {  }
+    weight: 1
+    region: content
+  field_short_title:
+    type: string
+    label: hidden
+    settings:
+      link_to_entity: false
+    third_party_settings: {  }
+    weight: 0
+    region: content
+hidden:
+  field_main_image: true
+  field_main_image_caption: true
+  field_news_groups: true
+  field_news_item_links_link: true
+  field_news_item_links_title: true
+  field_news_item_tags: true
+  field_news_neighbourhoods: true
+  field_radioactivity: true
+  langcode: true
+  links: true
+  published_at: true
+  search_api_excerpt: true
+  toc_enabled: true
