+++ conf/cmi/core.extension.yml
@@ -44,6 +44,7 @@ module:
   hdbt_admin_tools: 0
   health_check: 0
   helfi_alt_lang_fallback: 0
+  helfi_annif: 0
   helfi_api_base: 0
   helfi_azure_fs: 0
   helfi_base_content: 0
