+++ conf/cmi/search_api.index.recommendations.yml
@@ -1,0 +1,81 @@
+uuid: fcb3c79a-9d30-4639-9996-630999d0db40
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.node.field_lead_in
+    - search_api.server.news
+  module:
+    - helfi_annif
+    - node
+    - search_api
+    - helfi_etusivu
+    - helfi_react_search
+id: recommendations
+name: recommendations
+description: ''
+read_only: false
+field_settings:
+  ai_keywords:
+    label: 'AI Keywords'
+    property_path: ai_keywords
+    type: ai_keyword
+  entity_type:
+    label: 'Entity type'
+    property_path: search_api_entity_type
+    type: string
+  field_lead_in:
+    label: Lead
+    datasource_id: 'entity:node'
+    property_path: field_lead_in
+    type: noindex
+    dependencies:
+      config:
+        - field.storage.node.field_lead_in
+  title:
+    label: Title
+    datasource_id: 'entity:node'
+    property_path: title
+    type: noindex
+    dependencies:
+      module:
+        - node
+  uuid:
+    label: UUID
+    datasource_id: 'entity:node'
+    property_path: uuid
+    type: string
+    dependencies:
+      module:
+        - node
+datasource_settings:
+  'entity:node':
+    bundles:
+      default: false
+      selected:
+        - news_item
+    languages:
+      default: true
+      selected: {  }
+processor_settings:
+  add_url: {  }
+  aggregated_field: {  }
+  custom_value: {  }
+  district_image_absolute_url: {  }
+  entity_status: {  }
+  entity_type: {  }
+  keyword_generator: {  }
+  language_with_fallback: {  }
+  main_image_url: {  }
+  project_execution_schedule: {  }
+  project_image_absolute_url: {  }
+  project_plan_schedule: {  }
+  rendered_item: {  }
+tracker_settings:
+  default:
+    indexing_order: fifo
+options:
+  cron_limit: 50
+  index_directly: false
+  track_changes_in_references: true
+server: news
