+++ public/modules/custom/helfi_annif/helfi_annif.module
@@ -11,18 +11,20 @@ use Drupal\Core\Entity\EntityInterface;
 use Drupal\helfi_annif\TextConverter\Document;
 
 /**
- * Implements hook_ENTITY_TYPE_text_conversion_alter().
- */
-function helfi_annif_text_conversion_alter(Document &$document, EntityInterface $entity) {
-  // Strip nodes that contain classes 'is-hidden' class.
-  // @todo Move to platform config?
-  $document->stripNodes("//*[contains(@class, 'is-hidden')]");
-}
-
-/**
  * Implements hook_elasticsearch_connector_supported_data_types_alter().
  */
 function helfi_annif_elasticsearch_connector_supported_data_types_alter(array &$data_types) {
   $data_types[] = 'noindex';
   $data_types[] = 'ai_keyword';
+}
+
+/**
+ * Implements hook_text_conversion_alter().
+ */
+function helfi_annif_text_conversion_alter(Document &$document, EntityInterface $entity) : void {
+  // @todo Move to platform config?
+  // Strip nodes that are not displayed.
+  $document
+    ->stripNodes("//*[contains(@class, 'is-hidden')]")
+    ->stripNodes("//*[contains(@class, 'visually-hidden')]");
 }
