+++ public/modules/custom/helfi_annif/src/TextConverter/TextConverterManager.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Drupal\helfi_annif\TextConverter;
 
 use Drupal\Core\Entity\EntityInterface;
