+++ public/modules/custom/helfi_etusivu/helfi_etusivu.module
@@ -11,6 +11,7 @@ use Drupal\Core\Entity\ContentEntityInterface;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Language\LanguageInterface;
+use Drupal\helfi_annif\TextConverter\Document;
 use Drupal\helfi_platform_config\DTO\ParagraphTypeCollection;
 use Drupal\views\Plugin\views\query\QueryPluginBase;
 use Drupal\views\Plugin\views\query\Sql;
@@ -220,4 +221,16 @@ function helfi_etusivu_preprocess_page(&$variables): void {
 
   // Attach styles for the ordered news list and ordered news articles views.
   $variables['#attached']['library'][] = 'helfi_etusivu/views-styles';
+}
+
+/**
+ * Implements hook_ENTITY_TYPE_text_conversion_alter().
+ */
+function helfi_etusivu_node_text_conversion_alter(Document &$document, EntityInterface $entity) : void {
+  assert($entity instanceof \Drupal\node\NodeInterface);
+
+  if (in_array($entity->bundle(), ['news_item', 'news_article'])) {
+    $document
+      ->stripNodes("//*[contains(@class, 'content-date')]");
+  }
 }
